Remove deprecated user APIs
This commit removes generic user APIs are no longer used in production.

These have been supersded by CAS1 specific endpoints added via 3f48e89e833518dc971260cb2432068580fe789d

This commit also moves some CAS1 specific user header tests into the `Cas1UsersTest` class
