+++ src/main/resources/static/api.yml
@@ -3713,107 +3713,6 @@ paths:
           $ref: '_shared.yml#/components/responses/403Response'
         500:
           $ref: '_shared.yml#/components/responses/500Response'
-  /users/{id}:
-    get:
-      summary: Get information about a specific user
-      operationId: usersIdGet
-      parameters:
-        - in: path
-          name: id
-          required: true
-          description: Id of the user
-          schema:
-            type: string
-            format: uuid
-        - name: X-Service-Name
-          in: header
-          required: true
-          description: If given, only users for this service will be returned
-          schema:
-            $ref: '_shared.yml#/components/schemas/ServiceName'
-      responses:
-        200:
-          description: successfully retrieved information on user
-          content:
-            'application/json':
-              schema:
-                $ref: '_shared.yml#/components/schemas/User'
-        401:
-          $ref: '_shared.yml#/components/responses/401Response'
-        403:
-          $ref: '_shared.yml#/components/responses/403Response'
-        500:
-          $ref: '_shared.yml#/components/responses/500Response'
-    put:
-      summary: Update information about a specific user's roles and qualifications. Deprecated. Instead use PUT /cas1/user/{userid}
-      operationId: usersIdPut
-      deprecated: true
-      parameters:
-        - name: X-Service-Name
-          in: header
-          required: true
-          description: If given, only users for this service will be returned
-          schema:
-            $ref: '_shared.yml#/components/schemas/ServiceName'
-        - in: path
-          name: id
-          required: true
-          description: Id of the user
-          schema:
-            type: string
-            format: uuid
-      requestBody:
-        content:
-          'application/json':
-            schema:
-              $ref: '_shared.yml#/components/schemas/UserRolesAndQualifications'
-        required: true
-      responses:
-        200:
-          description: Successfully added information about user roles and qualifications
-          content:
-            'application/json':
-              schema:
-                $ref: '_shared.yml#/components/schemas/User'
-        401:
-          $ref: '_shared.yml#/components/responses/401Response'
-        403:
-          $ref: '_shared.yml#/components/responses/403Response'
-        500:
-          $ref: '_shared.yml#/components/responses/500Response'
-    delete:
-      summary: Deletes the user. Deprecated. Instead use DELETE /cas1/user/{userid}
-      operationId: usersIdDelete
-      deprecated: true
-      parameters:
-        - name: id
-          in: path
-          description: ID of the user
-          required: true
-          schema:
-            type: string
-            format: uuid
-        - name: X-Service-Name
-          in: header
-          required: true
-          description: Only users for this service will be allowed to delete a user
-          schema:
-            $ref: '_shared.yml#/components/schemas/ServiceName'
-      responses:
-        200:
-          description: successful operation
-        400:
-          description: invalid params
-          content:
-            'application/problem+json':
-              schema:
-                $ref: '_shared.yml#/components/schemas/ValidationError'
-        401:
-          $ref: '_shared.yml#/components/responses/401Response'
-        403:
-          $ref: '_shared.yml#/components/responses/403Response'
-        500:
-          $ref: '_shared.yml#/components/responses/500Response'
   /users/summary:
     get:
       tags:
