+++ src/test/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/integration/UsersTest.kt
@@ -1,321 +1,23 @@
 package uk.gov.justice.digital.hmpps.approvedpremisesapi.integration
 
-import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.EnumSource
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.data.repository.findByIdOrNull
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ApArea
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ApprovedPremisesUser
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ApprovedPremisesUserRole
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ProbationDeliveryUnit
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ProbationRegion
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ServiceName
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.TemporaryAccommodationUser
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.UserRolesAndQualifications
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.StaffDetailFactory
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.TeamFactoryDeliusContext
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.givens.givenACas1CruManagementArea
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.givens.givenAProbationRegion
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.givens.givenAUser
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.givens.givenAnApArea
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.httpmocks.apDeliusContextAddStaffDetailResponse
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.UserPermission
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.UserQualification
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.UserRole
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.deliuscontext.Borough
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.deliuscontext.PersonName
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.transformer.UserTransformer
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.util.randomStringMultiCaseWithNumbers
-import java.util.UUID
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.UserQualification as APIUserQualification
 
 class UsersTest : InitialiseDatabasePerClassTestBase() {
   @Autowired
   lateinit var userTransformer: UserTransformer
-
-  val id: UUID = UUID.fromString("aff9a4dc-e208-4e4b-abe6-99aff7f6af8a")
-
-  @Test
-  fun `Getting a user without a JWT returns 401`() {
-    webTestClient.get()
-      .uri("/users/$id")
-      .exchange()
-      .expectStatus()
-      .isUnauthorized
-  }
-
-  @Test
-  fun `Getting a user with a non-Delius JWT returns 403`() {
-    val jwt = jwtAuthHelper.createClientCredentialsJwt(
-      username = "username",
-      authSource = "other-auth-source",
-    )
-
-    webTestClient.get()
-      .uri("/users/$id")
-      .header("Authorization", "Bearer $jwt")
-      .exchange()
-      .expectStatus()
-      .isForbidden
-  }
-
-  @Test
-  fun `Getting a user with a Nomis JWT returns 403`() {
-    val jwt = jwtAuthHelper.createClientCredentialsJwt(
-      username = "username",
-      authSource = "nomis",
-    )
-
-    webTestClient.get()
-      .uri("/users/$id")
-      .header("Authorization", "Bearer $jwt")
-      .exchange()
-      .expectStatus()
-      .isForbidden
-  }
-
-  @Test
-  fun `Getting a user with the POM role returns 403`() {
-    val jwt = jwtAuthHelper.createClientCredentialsJwt(
-      username = "username",
-      authSource = "nomis",
-      roles = listOf("ROLE_POM"),
-    )
-
-    webTestClient.get()
-      .uri("/users/$id")
-      .header("Authorization", "Bearer $jwt")
-      .exchange()
-      .expectStatus()
-      .isForbidden
-  }
-
-  @Test
-  fun `Getting a user with no X-Service-Name header returns 400`() {
-    val deliusUsername = "JimJimmerson"
-    val forename = "Jim"
-    val surname = "Jimmerson"
-    val name = "$forename $surname"
-    val email = "foo@bar.com"
-    val telephoneNumber = "123445677"
-
-    val jwt = jwtAuthHelper.createAuthorizationCodeJwt(
-      subject = deliusUsername,
-      authSource = "delius",
-      roles = listOf("ROLE_PROBATION"),
-    )
-
-    mockClientCredentialsJwtRequest("username", listOf("ROLE_COMMUNITY"), authSource = "delius")
-
-    webTestClient.get()
-      .uri("/users/$id")
-      .header("Authorization", "Bearer $jwt")
-      .exchange()
-      .expectStatus()
-      .is4xxClientError
-      .expectBody()
-      .jsonPath("title").isEqualTo("Bad Request")
-      .jsonPath("detail").isEqualTo("Missing required header X-Service-Name")
-  }
-
-  @Test
-  fun `Getting an Approved Premises user returns OK with correct body`() {
-    val deliusUsername = "JimJimmerson"
-    val forename = "Jim"
-    val surname = "Jimmerson"
-    val name = "$forename $surname"
-    val email = "foo@bar.com"
-    val telephoneNumber = "123445677"
-
-    val jwt = jwtAuthHelper.createAuthorizationCodeJwt(
-      subject = deliusUsername,
-      authSource = "delius",
-      roles = listOf("ROLE_PROBATION"),
-    )
-
-    val region = givenAProbationRegion()
-
-    userEntityFactory.produceAndPersist {
-      withId(id)
-      withDeliusUsername(deliusUsername)
-      withName(name)
-      withEmail(email)
-      withTelephoneNumber(telephoneNumber)
-      withYieldedProbationRegion { region }
-    }
-
-    apDeliusContextAddStaffDetailResponse(
-      StaffDetailFactory.staffDetail(
-        name = PersonName(forename, surname),
-        deliusUsername = deliusUsername,
-        email = email,
-        telephoneNumber = telephoneNumber,
-      ),
-    )
-
-    mockClientCredentialsJwtRequest("username", listOf("ROLE_COMMUNITY"), authSource = "delius")
-
-    val result = webTestClient.get()
-      .uri("/users/$id")
-      .header("Authorization", "Bearer $jwt")
-      .header("X-Service-Name", ServiceName.approvedPremises.value)
-      .exchange()
-      .expectStatus()
-      .isOk
-      .returnResult(ApprovedPremisesUser::class.java)
-      .responseBody
-      .blockFirst()
-
-    assertThat(result.id).isEqualTo(id)
-    assertThat(result.region).isEqualTo(ProbationRegion(region.id, region.name))
-    assertThat(result.deliusUsername).isEqualTo(deliusUsername)
-    assertThat(result.name).isEqualTo(name)
-    assertThat(result.email).isEqualTo(email)
-    assertThat(result.telephoneNumber).isEqualTo(telephoneNumber)
-    assertThat(result.roles).isEqualTo(emptyList<ApprovedPremisesUserRole>())
-    assertThat(result.qualifications).isEqualTo(emptyList<UserQualification>())
-    assertThat(result.service).isEqualTo("CAS1")
-    assertThat(result.isActive).isEqualTo(true)
-    assertThat(result.apArea).isEqualTo(ApArea(region.apArea!!.id, region.apArea!!.identifier, region.apArea!!.name))
-    assertThat(result.permissions).isEqualTo(emptyList<UserPermission>())
-    assertThat(result.version).isNotZero()
-  }
-
-  @Test
-  fun `Getting a Temporary Accommodation user returns OK with correct body`() {
-    val deliusUsername = "JimJimmerson"
-    val forename = "Jim"
-    val surname = "Jimmerson"
-    val name = "$forename $surname"
-    val email = "foo@bar.com"
-    val telephoneNumber = "123445677"
-    val borough = Borough(
-      code = randomStringMultiCaseWithNumbers(10),
-      description = randomStringMultiCaseWithNumbers(10),
-    )
-
-    val jwt = jwtAuthHelper.createAuthorizationCodeJwt(
-      subject = deliusUsername,
-      authSource = "delius",
-      roles = listOf("ROLE_PROBATION"),
-    )
-
-    val region = givenAProbationRegion()
-
-    val probationDeliveryUnit = probationDeliveryUnitFactory.produceAndPersist {
-      withDeliusCode(borough.code)
-      withName(borough.description)
-      withProbationRegion(region)
-    }
-
-    userEntityFactory.produceAndPersist {
-      withId(id)
-      withDeliusUsername(deliusUsername)
-      withName(name)
-      withEmail(email)
-      withTelephoneNumber(telephoneNumber)
-      withYieldedProbationRegion { region }
-      withProbationDeliveryUnit { probationDeliveryUnit }
-    }
-
-    apDeliusContextAddStaffDetailResponse(
-      StaffDetailFactory.staffDetail(
-        name = PersonName(forename, surname),
-        deliusUsername = deliusUsername,
-        email = email,
-        telephoneNumber = telephoneNumber,
-        teams = listOf(TeamFactoryDeliusContext.team(borough = borough)),
-      ),
-    )
-
-    mockClientCredentialsJwtRequest("username", listOf("ROLE_COMMUNITY"), authSource = "delius")
-
-    webTestClient.get()
-      .uri("/users/$id")
-      .header("Authorization", "Bearer $jwt")
-      .header("X-Service-Name", ServiceName.temporaryAccommodation.value)
-      .exchange()
-      .expectStatus()
-      .isOk
-      .expectBody()
-      .json(
-        objectMapper.writeValueAsString(
-          TemporaryAccommodationUser(
-            id = id,
-            region = ProbationRegion(region.id, region.name),
-            probationDeliveryUnit = ProbationDeliveryUnit(probationDeliveryUnit.id, probationDeliveryUnit.name),
-            deliusUsername = deliusUsername,
-            name = name,
-            email = email,
-            telephoneNumber = telephoneNumber,
-            roles = emptyList(),
-            service = "CAS3",
-            isActive = true,
-          ),
-        ),
-      )
-  }
-
-  @Nested
-  inner class UserHeaders {
-
-    @Test
-    fun `Getting an Approved Premises user with no roles returns x cas user headers`() {
-      val (user, jwt) = givenAUser()
-
-      val headers = webTestClient.get()
-        .uri("/users/${user.id}")
-        .header("Authorization", "Bearer $jwt")
-        .header("X-Service-Name", ServiceName.approvedPremises.value)
-        .exchange()
-        .expectStatus()
-        .isOk
-        .returnResult(ApprovedPremisesUser::class.java)
-        .responseHeaders
-
-      assertThat(headers["X-CAS-User-ID"]!![0]).isEqualTo(user.id.toString())
-      assertThat(headers["X-CAS-User-Version"]!![0]).isEqualTo("993")
-    }
-
-    @Test
-    fun `Getting an Approved Premises user with roles returns x cas user headers`() {
-      val (user, jwt) = givenAUser(roles = listOf(UserRole.CAS1_REPORT_VIEWER))
-
-      val headers = webTestClient.get()
-        .uri("/users/${user.id}")
-        .header("Authorization", "Bearer $jwt")
-        .header("X-Service-Name", ServiceName.approvedPremises.value)
-        .exchange()
-        .expectStatus()
-        .isOk
-        .returnResult(ApprovedPremisesUser::class.java)
-        .responseHeaders
-
-      assertThat(headers["X-CAS-User-ID"]!![0]).isEqualTo(user.id.toString())
-      assertThat(headers["X-CAS-User-Version"]!![0]).isEqualTo("1173003502")
-    }
-
-    @Test
-    fun `Getting a non Approved Premises user does not return x cas user headers`() {
-      val (user, jwt) = givenAUser()
-
-      val headers = webTestClient.get()
-        .uri("/users/${user.id}")
-        .header("Authorization", "Bearer $jwt")
-        .header("X-Service-Name", ServiceName.temporaryAccommodation.value)
-        .exchange()
-        .expectStatus()
-        .isOk
-        .returnResult(ApprovedPremisesUser::class.java)
-        .responseHeaders
-
-      assertThat(headers.containsKey("X-CAS-User-ID")).isFalse()
-      assertThat(headers.containsKey("X-CAS-User-Version")).isFalse()
-    }
-  }
 
   @Nested
   inner class GetUsers {
@@ -1158,238 +860,6 @@ class UsersTest : InitialiseDatabasePerClassTestBase() {
           .jsonPath("title").isEqualTo("Not Found")
           .jsonPath("status").isEqualTo(404)
           .jsonPath("detail").isEqualTo("No user with username of noone could be found")
-      }
-    }
-  }
-
-  @Nested
-  inner class UpdateUser {
-    @ParameterizedTest
-    @EnumSource(value = UserRole::class, names = ["CAS1_WORKFLOW_MANAGER", "CAS1_JANITOR", "CAS1_USER_MANAGER"])
-    fun `Updating a user update with X-Service-Name other than approved-premises is forbidden`(role: UserRole) {
-      givenAUser(roles = listOf(role)) { _, jwt ->
-        val id = UUID.randomUUID()
-        webTestClient.put()
-          .uri("/users/$id")
-          .header("Authorization", "Bearer $jwt")
-          .header("X-Service-Name", ServiceName.temporaryAccommodation.value)
-          .header("Content-Type", "application/json")
-          .bodyValue(
-            UserRolesAndQualifications(
-              listOf<ApprovedPremisesUserRole>(),
-              listOf<APIUserQualification>(),
-            ),
-          )
-          .exchange()
-          .expectStatus()
-          .isForbidden
-      }
-    }
-
-    @ParameterizedTest
-    @EnumSource(value = UserRole::class, names = ["CAS1_FUTURE_MANAGER", "CAS1_MATCHER", "CAS1_JANITOR", "CAS1_USER_MANAGER"], mode = EnumSource.Mode.EXCLUDE)
-    fun `Updating a user without an approved role is forbidden`(role: UserRole) {
-      givenAUser(roles = listOf(role)) { _, jwt ->
-        val id = UUID.randomUUID()
-        webTestClient.put()
-          .uri("/users/$id")
-          .header("Authorization", "Bearer $jwt")
-          .header("X-Service-Name", ServiceName.temporaryAccommodation.value)
-          .header("Content-Type", "application/json")
-          .bodyValue(
-            UserRolesAndQualifications(
-              listOf<ApprovedPremisesUserRole>(),
-              listOf<APIUserQualification>(),
-            ),
-          )
-          .exchange()
-          .expectStatus()
-          .isForbidden
-      }
-    }
-
-    @Test
-    fun `Updating a users with no internal role (aka the Applicant pseudo-role) is forbidden`() {
-      givenAUser { _, jwt ->
-        val id = UUID.randomUUID()
-        webTestClient.put()
-          .uri("/users/$id")
-          .header("Authorization", "Bearer $jwt")
-          .header("X-Service-Name", ServiceName.approvedPremises.value)
-          .bodyValue(
-            UserRolesAndQualifications(
-              listOf<ApprovedPremisesUserRole>(),
-              listOf<APIUserQualification>(),
-            ),
-          )
-          .exchange()
-          .expectStatus()
-          .isForbidden
-      }
-    }
-
-    @ParameterizedTest
-    @EnumSource(value = UserRole::class, names = ["CAS1_JANITOR", "CAS1_USER_MANAGER"])
-    fun `Updating a user returns OK with correct body when user has an approved role`(role: UserRole) {
-      val id = UUID.randomUUID()
-      val qualifications = listOf(APIUserQualification.emergency, APIUserQualification.pipe)
-      val roles = listOf(
-        ApprovedPremisesUserRole.assessor,
-        ApprovedPremisesUserRole.reportViewer,
-        ApprovedPremisesUserRole.excludedFromAssessAllocation,
-        ApprovedPremisesUserRole.excludedFromMatchAllocation,
-        ApprovedPremisesUserRole.excludedFromPlacementApplicationAllocation,
-      )
-      val region = givenAProbationRegion()
-
-      givenAUser(
-        id = id,
-        isActive = false,
-        probationRegion = region,
-      )
-
-      givenAUser(roles = listOf(role)) { _, jwt ->
-        webTestClient.put()
-          .uri("/users/$id")
-          .header("Authorization", "Bearer $jwt")
-          .header("X-Service-Name", ServiceName.approvedPremises.value)
-          .bodyValue(
-            UserRolesAndQualifications(
-              roles = roles,
-              qualifications = qualifications,
-            ),
-          )
-          .exchange()
-          .expectStatus()
-          .isOk
-          .expectBody()
-          .jsonPath(".qualifications").isArray
-          .jsonPath(".qualifications[0]").isEqualTo("emergency")
-          .jsonPath(".roles").isArray
-          .jsonPath(".roles[0]").isEqualTo(ApprovedPremisesUserRole.assessor.value)
-          .jsonPath(".roles[1]").isEqualTo(ApprovedPremisesUserRole.reportViewer.value)
-          .jsonPath(".roles[2]").isEqualTo(ApprovedPremisesUserRole.excludedFromAssessAllocation.value)
-          .jsonPath(".roles[3]").isEqualTo(ApprovedPremisesUserRole.excludedFromMatchAllocation.value)
-          .jsonPath(".roles[4]").isEqualTo(ApprovedPremisesUserRole.excludedFromPlacementApplicationAllocation.value)
-          .jsonPath(".isActive").isEqualTo(true)
-      }
-    }
-  }
-
-  @Nested
-  inner class DeleteUser {
-    @ParameterizedTest
-    @EnumSource(value = UserRole::class, names = ["CAS1_WORKFLOW_MANAGER", "CAS1_JANITOR", "CAS1_USER_MANAGER"])
-    fun `Deleting a user with X-Service-Name other than approved-premises is forbidden`(role: UserRole) {
-      givenAUser(roles = listOf(role)) { _, jwt ->
-        val id = UUID.randomUUID()
-        webTestClient.delete()
-          .uri("/users/$id")
-          .header("Authorization", "Bearer $jwt")
-          .header("X-Service-Name", ServiceName.temporaryAccommodation.value)
-          .header("Content-Type", "application/json")
-          .exchange()
-          .expectStatus()
-          .isForbidden
-      }
-    }
-
-    @ParameterizedTest
-    @EnumSource(value = UserRole::class, names = ["CAS1_WORKFLOW_MANAGER", "CAS1_JANITOR", "CAS1_USER_MANAGER"], mode = EnumSource.Mode.EXCLUDE)
-    fun `Deleting a user with an unapproved role is forbidden`(role: UserRole) {
-      givenAUser(roles = listOf(role)) { _, jwt ->
-        val id = UUID.randomUUID()
-        webTestClient.delete()
-          .uri("/users/$id")
-          .header("Authorization", "Bearer $jwt")
-          .header("X-Service-Name", ServiceName.approvedPremises.value)
-          .header("Content-Type", "application/json")
-          .exchange()
-          .expectStatus()
-          .isForbidden
-      }
-    }
-
-    @Test
-    fun `Deleting a user with no internal role (aka the Applicant pseudo-role) is forbidden`() {
-      givenAUser { _, jwt ->
-        val id = UUID.randomUUID()
-        webTestClient.delete()
-          .uri("/users/$id")
-          .header("Authorization", "Bearer $jwt")
-          .header("X-Service-Name", ServiceName.approvedPremises.value)
-          .exchange()
-          .expectStatus()
-          .isForbidden
-      }
-    }
-
-    @ParameterizedTest
-    @EnumSource(value = UserRole::class, names = ["CAS1_WORKFLOW_MANAGER", "CAS1_JANITOR", "CAS1_USER_MANAGER"])
-    fun `Deleting a user with no X-Service-Name is forbidden`() {
-      givenAUser { _, jwt ->
-        val id = UUID.randomUUID()
-        webTestClient.delete()
-          .uri("/users/$id")
-          .header("Authorization", "Bearer $jwt")
-          .exchange()
-          .expectStatus()
-          .is4xxClientError
-          .expectBody()
-          .jsonPath("title").isEqualTo("Bad Request")
-          .jsonPath("detail").isEqualTo("Missing required header X-Service-Name")
-      }
-    }
-
-    @Test
-    fun `Deleting a user with a non-Delius JWT returns 403`() {
-      val jwt = jwtAuthHelper.createClientCredentialsJwt(
-        username = "username",
-        authSource = "nomis",
-      )
-
-      webTestClient.delete()
-        .uri("/users/$id")
-        .header("Authorization", "Bearer $jwt")
-        .header("X-Service-Name", ServiceName.approvedPremises.value)
-        .exchange()
-        .expectStatus()
-        .isForbidden
-    }
-
-    @ParameterizedTest
-    @EnumSource(value = UserRole::class, names = ["CAS1_JANITOR", "CAS1_USER_MANAGER"])
-    fun `Deleting a user with an approved role deletes successfully`(role: UserRole) {
-      userEntityFactory.produceAndPersist {
-        withId(id)
-        withYieldedProbationRegion { givenAProbationRegion() }
-      }
-
-      givenAUser(roles = listOf(role)) { _, jwt ->
-        webTestClient.delete()
-          .uri("/users/$id")
-          .header("Authorization", "Bearer $jwt")
-          .header("X-Service-Name", ServiceName.approvedPremises.value)
-          .exchange()
-          .expectStatus()
-          .isOk
-
-        val userFromDatabase = userRepository.findByIdOrNull(id)
-        assertThat(userFromDatabase?.isActive).isEqualTo(false)
-      }
-    }
-
-    @ParameterizedTest
-    @EnumSource(value = UserRole::class, names = ["CAS1_WORKFLOW_MANAGER", "CAS1_JANITOR", "CAS1_USER_MANAGER"], mode = EnumSource.Mode.EXCLUDE)
-    fun `Deleting a user without an approved role is forbidden `(role: UserRole) {
-      givenAUser(roles = listOf(role)) { _, jwt ->
-        webTestClient.delete()
-          .uri("/users/$id")
-          .header("Authorization", "Bearer $jwt")
-          .header("X-Service-Name", ServiceName.approvedPremises.value)
-          .exchange()
-          .expectStatus()
-          .isForbidden
       }
     }
   }
